Prevent items with placeholders from exceeding bank capacity
When the bank is full, items which have placeholders but are not present in the bank (quantity 0) will still be added via combat loot or skill rewards, ignoring the bank's capacity. This change adds a patch to the `Bank.addItem()` method to handle placeholders as a special case, preventing those items from exceeding the bank limit.

----

The `Rewards.giveRewards()` method used by most skills to determine whether a gathered or crafted item can be added to the bank does not call `Bank.hasItem()` or `Bank.willItemsFit()` to check for empty slots, but instead calls `Bank.addItem()` directly.
However, any drop sent to `addItem` will be assumed to fit as long as the item is defined in the bank items map (since it assumes the existing quantity can't be 0).
The same is true for the `CombatLoot.lootItem()` and `CombatLoot.lootAll()` methods called when receiving combat loot.

This change replaces the `addItem()` method, but passes any item through to be handled normally if it satisfies one of:
- `!isPlaceholder(bankItem)` – Has non-zero quantity or is not present in the bank
- `useSlots` – The 'Use slots' setting is on, so placeholders count towards bank space anyway
- `this.occupiedSlots < this.maximumSlots` – There's at least one bank slot open
- `ignoreSpace` – Parameter set to ignore bank capacity

Otherwise, the original method is skipped and a `BankFull` notification is fired, along with potentially logging the item as lost (behavior copied from the original `addItem`).